+++ coast/mtfoundation/Threads.ipp
@@ -9,23 +9,14 @@
 #ifndef _Threads_IPP
 #define _Threads_IPP
 
-//--- project modules used -----------------------------------------------------
-
-//--- standard modules used ----------------------------------------------------
-
-//--- c-modules used -----------------------------------------------------------
-
-//:Try to set Working state
-template< class WorkerParamType >
-bool Thread::SetWorking(WorkerParamType workerArgs)
-{
-	StatTrace(Thread.SetWorking, "ThrdId: " << (long)GetId() << " CallId: " << MyId(), Coast::Storage::Current());
+template<class WorkerParamType>
+bool Thread::SetWorking(WorkerParamType workerArgs) {
+	StatTrace(Thread.SetWorking, "ThrdId: " << GetId() << " CallId: " << MyId(), Coast::Storage::Current());
 	return SetRunningState(eWorking, workerArgs);
 }
 
-template< class WorkerParamType >
-bool Thread::SetRunningState(ERunningState state, WorkerParamType args)
-{
+template<class WorkerParamType>
+bool Thread::SetRunningState(ERunningState state, WorkerParamType args) {
 	LockUnlockEntry me(fStateMutex);
 
 	// allocate things used before and after call to CallRunningStateHooks() on Coast::Storage::Global() because Allocator could be refreshed during call
@@ -43,10 +34,10 @@ bool Thread::SetRunningState(ERunningState state, WorkerParamType args)
 	{
 		// scoping to force compiler not to move around automatic variables which could make strange things happen
 		Anything anyEvt;
-		anyEvt["ThreadId"] = (long)GetId();
-		anyEvt["RunningState"]["Old"] = (long)oldState;
-		anyEvt["RunningState"]["New"] = (long)state;
-		if ( !args.IsNull() ) {
+		anyEvt["ThreadId"] = GetId();
+		anyEvt["RunningState"]["Old"] = (long) oldState;
+		anyEvt["RunningState"]["New"] = (long) state;
+		if (!args.IsNull()) {
 			anyEvt["Args"] = args.DeepClone();
 		}
 		BroadCastEvent(anyEvt);
